SSCS-7687 remove dodgy before
