+++ src/e2e/java/uk/gov/hmcts/reform/sscs/functional/EvidenceShareFunctionalTest.java
@@ -1,12 +1,13 @@
 package uk.gov.hmcts.reform.sscs.functional;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.VALID_APPEAL_CREATED;
 
 import io.github.artsok.RepeatedIfExceptionsTest;
 import java.time.LocalDate;
 import java.util.List;
 
+import org.junit.Test;
 import org.junit.jupiter.api.BeforeEach;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
@@ -44,5 +45,43 @@ public class EvidenceShareFunctionalTest extends AbstractFunctionalTest {
         assertEquals("withDwp", caseDetails.getState());
         assertEquals(LocalDate.now().toString(), caseData.getDateSentToDwp());
         assertEquals(LocalDate.now().toString(), caseData.getDateCaseSentToGaps());
+    }
+
+    @RepeatedIfExceptionsTest(repeats = 3, suspend = 5000)
+    public void processAnAppealWithNoValidMrnDate_shouldNoTBeSentToDwpAndShouldBeUpdatedToFlagError() throws Exception {
+
+        createCaseWithValidAppealState(VALID_APPEAL_CREATED);
+
+        String json = getJson(VALID_APPEAL_CREATED.getCcdType());
+        json = json.replace("CASE_ID_TO_BE_REPLACED", ccdCaseId);
+        json = json.replace("MRN_DATE_TO_BE_REPLACED", "");
+
+        simulateCcdCallback(json);
+        SscsCaseDetails caseDetails = findCaseById(ccdCaseId);
+
+        assertNull(caseDetails.getData().getSscsDocument());
+        assertEquals("validAppeal", caseDetails.getState());
+        assertEquals("failedSending", caseDetails.getData().getHmctsDwpState());
+    }
+
+    @Test
+    public void processAnAppealWithLateMrn_shouldGenerateADl16AndAddToCcdAndUpdateState() throws Exception {
+        createCaseWithValidAppealState(VALID_APPEAL_CREATED);
+
+        String json = getJson(VALID_APPEAL_CREATED.getCcdType());
+        json = json.replace("CASE_ID_TO_BE_REPLACED", ccdCaseId);
+        json = json.replace("MRN_DATE_TO_BE_REPLACED", LocalDate.now().minusDays(31).toString());
+
+        simulateCcdCallback(json);
+        SscsCaseDetails caseDetails = findCaseById(ccdCaseId);
+        SscsCaseData caseData = caseDetails.getData();
+
+        List<SscsDocument> docs = caseData.getSscsDocument();
+
+        assertNotNull("docs is not null", docs);
+        assertEquals(1, docs.size());
+        assertEquals("dl16-" + ccdCaseId + ".pdf", docs.get(0).getValue().getDocumentFileName());
+        assertEquals("withDwp", caseDetails.getState());
+        assertEquals(LocalDate.now().toString(), caseData.getDateSentToDwp());
     }
 }
